processed data without weird data breaks
Here is the processed data ignoring all weird instances of transcripts being on 2 lines. As a note, in data_transcript_fully_processed.csv, processed_transcript refers to the column where I removed stopwords and lemmatized. I could not get the version with no contractions to also upload in the same file.